visionimg cleanup
