+++ selfdrive/ui/qt/home.cc
@@ -268,19 +268,6 @@ void GLWindow::wake() {
   }
 }
 
-GLuint visionimg_to_gl(const VisionImg *img, EGLImageKHR *pkhr, void **pph) {
-  unsigned int texture;
-  glGenTextures(1, &texture);
-  glBindTexture(GL_TEXTURE_2D, texture);
-  glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB, img->width, img->height, 0, GL_RGB, GL_UNSIGNED_BYTE, *pph);
-  glGenerateMipmap(GL_TEXTURE_2D);
-  *pkhr = (EGLImageKHR)1; // not NULL
-  return texture;
-}
-
-void visionimg_destroy_gl(EGLImageKHR khr, void *ph) {
-  // empty
-}
 
 FramebufferState* framebuffer_init(const char* name, int32_t layer, int alpha,
                                    int *out_w, int *out_h) {
